requirements.txt
pandas>=2.1
numpy>=1.26
scikit-learn>=1.3
joblib>=1.3
matplotlib>=3.8
kafka-python>=2.0.2
SQLAlchemy>=2.0
python-dotenv>=1.0

